Backend/requirements.txt
google-generativeai
Pillow
fastapi
uvicorn
pydantic
python-dotenv